requirements.txt
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.7.0
certifi==2023.11.17
cffi==1.16.0
click==8.1.7
colorama==0.4.6
Flask==3.0.0
h11==0.14.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
outcome==1.3.0.post0
pycparser==2.21
PyMySQL==1.1.0
PySocks==1.7.1
python-dotenv==1.0.0
selenium==4.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
trio==0.23.2
trio-websocket==0.11.1
types-beautifulsoup4==4.12.0.7
types-html5lib==1.1.11.15
urllib3==2.1.0
Werkzeug==3.0.1
wsproto==1.2.0
